+++ pygeoapi/processes/location_info_time_rasterstats.py
@@ -1,8 +1,9 @@
+# noqa: D100
 import logging
 import shapely
 
 from pygeoapi.process.base import BaseProcessor
-from .algorithms.location_info import get_location_info_time
+from .algorithms.rasterstats_algorithms import get_location_info_time
 from .algorithms.util import (get_crs_from_cog,
                               reproject,
                               url_exists,
